extract into a separate component
